Fix intermittent problems with dropping a native file in FF
Fixes #539

When dropping a native file in FF, under some conditions the `mousemove` event was firing when the file was dropped before the `drop` event fired. If the drop operation would fail.

To fix, the event handler is delayed so that the end drag will always run after the drop handler. This commit also switches to using the `mouseover` event so that the drag state is reset sooner if the original FF dragleave bug is hit, and the commit resets the `EnterLeaveCounter` when ending the drag operation via the `mouseover` event handler because the FF `dragleave` bug results in `this.entered` not being empty.