docs: move documentation examples into a separate package
This allows us to perform typescript validation on the examples outside of the Gatsby site, which uses babel and misses some issues.

Unit tests for the examples will eventually land here as well.

Ultimately, the text portion of these examples will go back into Gatsby as Markdown, including the functional example as a custom component.