fix: resolve two-backend issue
When the DragDropProvider is removed from the DOM, it was not emitting a message to the DragDropManager so that it could handle refcounts properly and break-down the HTML5Backend.

This change adds a useEffect hook invocation to fire off an event that will cause the DnDManager to destroy the backend.