Small wording change for Tutorial.md
I was going through the tutorial for React DnD and noticed that on line 39, it reads `are use function-components` instead of `are using function-components`.

Thank you so much for the in depth introduction to React DnD. I've gone through the whole tutorial now and have my first working example of the library that I'm planning to use a lot. Not only did it give great insight into how React DnD works, but you did an outstanding job of breaking down design decisions and linking to helpful articles to learn even more.