Fix isNodeInDocument for HTML5Backend when used in iframes
The isNodeInDocument function is used to check if a node was detached
during a drag event. It checks to make sure this `this.document.body` is
defined, but then checks if `document.body` contains the given node.

When used within an iframe, `document` in the global context may not be
the same as `this.document`. This commit updates the check to use
`this.document.body`.