Update useDrop.md, remove accept as a function (#3319)
Looks like this doc is incorrect, `accept` does not appear to take a function. When trying to use a function for `accept`, I see this error. `Uncaught Invariant Violation: Type can only be a string or a symbol.`

`DropTargetHookSpec` says accept is `TargetType`
https://github.com/react-dnd/react-dnd/blob/e8bd6436548d96f6d6594f763752f424c2e0834b/packages/react-dnd/src/hooks/types.ts#L97 Lists
`TargetType` is `Identifier | Identifier[] | null`
https://github.com/react-dnd/react-dnd/blob/e8bd6436548d96f6d6594f763752f424c2e0834b/packages/dnd-core/src/utils/matchesType.ts
And `Identifier` is `string | symbol`
https://github.com/react-dnd/react-dnd/blob/e8bd6436548d96f6d6594f763752f424c2e0834b/packages/dnd-core/src/interfaces.ts#L1